commit 3: [alter view as] statement
