+++ fe/src/main/cup/sql_parser.cup
@@ -616,9 +616,9 @@ alter_stmt ::=
         RESULT = new AlterTableStmt(tbl, clauses);
     :}
     | KW_ALTER KW_VIEW table_name:tbl
-    alter_table_clause_list:clauses
+    opt_col_with_comment_list:columns KW_AS query_stmt:view_def
     {:
-        RESULT = new AlterViewStmt(tbl, clauses);
+        RESULT = new AlterViewStmt(tbl, columns, view_def);
     :}
     | KW_ALTER KW_SYSTEM alter_system_clause:clause
     {:
@@ -780,6 +780,7 @@ alter_table_clause ::=
     {:
         RESULT = new ColumnRenameClause(colName, newColName);
     :}
+    | KW_MODIFY
     ;
 
 alter_system_clause ::=
